Added some entity handling stuff

+++ core/src/main/java/com/ktar5/slime/world/tiles/Drain.java
@@ -1,7 +1,8 @@
 package com.ktar5.slime.world.tiles;
 
+import com.ktar5.slime.engine.entities.Entity;
 import com.ktar5.slime.engine.util.Side;
-import com.ktar5.slime.player.JumpPlayer;
+import com.ktar5.slime.entities.player.JumpPlayer;
 import com.ktar5.slime.world.tiles.base.Rotation;
 import com.ktar5.slime.world.tiles.base.WholeTile;
 
@@ -17,16 +18,19 @@ public class Drain extends WholeTile {
     }
 
     @Override
-    public void onPlayerCross(JumpPlayer player) {
-        if(player.isSmall()){
-            player.kill();
-        }else{
-            player.setSmall(true);
+    public void onCross(Entity entity) {
+        if (entity.isPlayer()) {
+            JumpPlayer player = (JumpPlayer) entity;
+            if(player.isSmall()){
+                player.kill();
+            }else{
+                player.setSmall(true);
+            }
         }
     }
 
     @Override
-    public boolean canCrossThrough(JumpPlayer player, Side movement) {
+    public boolean canCrossThrough(Entity entity, Side movement) {
         return true;
     }
 }
